Add non-functional preferences window
Add a simple preferences window. None of the functionality is connected
yet to anything, but the main UI is there now.

The design took inspiration from Google's Material Design, especially
with respect to distances and metrics. Other apps with a search engine
or services selection were also briefly studied.

Key actions that a user wants to perform when opening this window are
* enabling and disabling search engines,
* adding search engines and configuring keywords, and
* deleting search engines that they feel are no longer of use to them.

The first two can be served well by the UI in the future, but for the
latter a "delete multiple" feature could be more useful.

However, selecting multiple items is not well supported by GTK, the
underlying toolkit.