Add settings schema for search engine details
This is the first step to storing search engine preferences with
GSettings. Each search engine will be identified by a hash of its name.

The user will be able to keep a list of search engines they want to use,
as well as their preference order. Name, query URL, and other details
can be stored for each search engine.

To store the settings, each search engine needs a unique identifier.
Ideally, this identifier is based on related metadata. It should only
use lowercase alphanumeric characters and ASCII hyphen ("-"), otherwise
it could not be used as a GSettings key.

Other popular search tools such as Krunner or the Chromium search
require each search engine to have a unique keyword which doubles as the
identifier. Doing this while using the keywords as a GSettings keys
would forbid non-ASCII keywords. Moreover, although keywords are a
useful feature, they are not *required* for basic functionality.
Therefore, the identifier is not based on a keyword.

Not only the search engine's identifier to store settings, but also its
name should be unique: If there were two search engines with the same
name, it could be confusing which item launches which search. Unique
names offer a good basis for GSettings keys.

Search engine names could be case-sensitive and contain non-alphanumeric
characters (e.g. spaces), so they cannot be used as a GSettings key
directly. Instead, a hash of the name is used.