Store settings in a simple JSON file instead of GSettings
Search engines are stored in a simple JSON file and will only be loaded
when the extension is activated.

This greatly simplifies the implementation of a customisable list of
search engines.

GSettings is not used because after a substantial amount of time I made
little progress in storing the settings with GSettings and a small
preferences window.